LPS-105024 Move types to API
