+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-api/build.gradle
@@ -3,5 +3,4 @@ dependencies {
 	compileOnly group: "org.json", name: "json", version: "20180813"
 	compileOnly group: "org.osgi", name: "org.osgi.annotation.versioning", version: "1.1.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	compileOnly project(":apps:portal-security-crypto:portal-security-crypto-generator-api")
 }
