+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-api/src/main/java/com/liferay/portal/security/crypto/generator/hashing/salt/VariableLengthSaltGenerator.java
@@ -12,25 +12,22 @@
  * details.
  */
 
-package com.liferay.portal.security.crypto.generator.spi.hashing;
+package com.liferay.portal.security.crypto.generator.hashing.salt;
 
-import com.liferay.portal.security.crypto.generator.spi.Generator;
-import com.liferay.portal.security.crypto.generator.spi.hashing.salt.SaltGenerator;
-
-import org.osgi.annotation.versioning.ProviderType;
+import com.liferay.portal.kernel.io.BigEndianCodec;
+import com.liferay.portal.kernel.security.SecureRandomUtil;
 
 /**
- * @author Arthur Chan
+ * @author Carlos Sierra Andrés
  */
-@ProviderType
-public interface HashGenerator extends Generator {
+public interface VariableLengthSaltGenerator extends SaltGenerator {
 
-	public SaltGenerator getSaltGenerator() throws Exception;
+	public default byte[] generateSalt(int size) {
+		byte[] saltBytes = new byte[size];
 
-	public byte[] hash(byte[] toBeHashed) throws Exception;
+		BigEndianCodec.putLong(saltBytes, 0, SecureRandomUtil.nextLong());
 
-	public HashGenerator withPepper(byte[] pepper) throws Exception;
-
-	public HashGenerator withSalt(byte[] salt) throws Exception;
+		return saltBytes;
+	}
 
 }
