+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-api/src/main/resources/com/liferay/portal/security/crypto/generator/packageinfo
