+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-spi/bnd.bnd
@@ -2,7 +2,5 @@ Bundle-Name: Liferay Portal Security Crypto Generator SPI
 Bundle-SymbolicName: com.liferay.portal.security.crypto.generator.spi
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.portal.security.crypto.generator.spi,\
 	com.liferay.portal.security.crypto.generator.spi.hashing,\
-	com.liferay.portal.security.crypto.generator.spi.hashing.factory,\
-	com.liferay.portal.security.crypto.generator.spi.hashing.salt
+	com.liferay.portal.security.crypto.generator.spi.hashing.factory
